passe les float en double

+++ unitesLib/src/unitesLib/Unite.java
@@ -2,28 +2,28 @@ package unitesLib;
 
 public class Unite {
 	private String nom;
-	private float coef; //coefficient multiplicateur
-	private float decal; // coefficient de decallage
+	private double coef; //coefficient multiplicateur
+	private double decal; // coefficient de decallage
 
 	//constructeur sans décalage
-	public Unite(String pNom, float pVal){
+	public Unite(String pNom, double pVal){
 		nom = pNom;
 		coef=pVal;
 		decal=0;
 	}
 	
 	//constructeur avec décallage
-	public Unite(String pNom, float pVal, float pDecal){
+	public Unite(String pNom, double pVal, double pDecal){
 		nom = pNom;
 		coef=pVal;
 		decal=pDecal;
 	}
 	
-	public float getDecal() {
+	public double getDecal() {
 		return decal;
 	}
 
-	public void setDecal(float decal) {
+	public void setDecal(double decal) {
 		this.decal = decal;
 	}
 	
@@ -35,11 +35,11 @@ public class Unite {
 		this.nom = nom;
 	}
 
-	public float getVal() {
+	public double getVal() {
 		return coef;
 	}
 
-	public void setVal(float val) {
+	public void setVal(double val) {
 		this.coef = val;
 	}
 }
